+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - SQSEventPublisher to publish events using AWS Simple Queue Service
+- SQSEvent & SQSEventBatch as specialized handlers for Simple Queue Service-backed lambdas
 
 ## [0.49.0] - 2018-08-23
 ### Added